requirements.txt
streamlit==1.37.1
opencv-python==4.10.0.84
Pillow==10.4.0
numpy==1.26.4
matplotlib==3.9.2
requests==2.32.3
